+++ src/components/routing/Routes.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import Contact from '../contact/Contact';
+import Projects from '../projects/Projects';
 import Footer from '../layout/Footer';
 
 const Routes = () => {
@@ -10,6 +11,9 @@ const Routes = () => {
             <Switch>
                 <Route exact path='/contact'>
                     <Contact />
+                </Route>
+                <Route exact path='/projects'>
+                    <Projects />
                 </Route>
             </Switch>
             <Footer />